+++ mobile/src/main/java/treehou/se/habit/ui/menu/NavigationDrawerFragment.java
@@ -4,7 +4,6 @@ package treehou.se.habit.ui.menu;
 import android.content.Context;
 import android.support.annotation.IntDef;
 import android.support.v7.app.ActionBarDrawerToggle;
-import android.support.v4.app.Fragment;
 import android.support.v4.view.GravityCompat;
 import android.support.v4.widget.DrawerLayout;
 import android.content.SharedPreferences;
@@ -12,29 +11,36 @@ import android.content.res.Configuration;
 import android.os.Bundle;
 import android.support.v7.app.ActionBar;
 import android.support.v7.app.AppCompatActivity;
+import android.support.v7.widget.DefaultItemAnimator;
+import android.support.v7.widget.LinearLayoutManager;
+import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.AdapterView;
-import android.widget.ArrayAdapter;
 import android.widget.ImageView;
-import android.widget.ListView;
 import android.widget.TextView;
+
+import com.trello.rxlifecycle.RxLifecycle;
+import com.trello.rxlifecycle.components.support.RxFragment;
 
 import java.util.ArrayList;
 import java.util.List;
 
 import javax.inject.Inject;
 
+import io.realm.Realm;
+import rx.android.schedulers.AndroidSchedulers;
+import se.treehou.ng.ohcommunicator.connector.models.OHSitemap;
 import treehou.se.habit.HabitApplication;
 import treehou.se.habit.R;
 import treehou.se.habit.module.ApplicationComponent;
+import treehou.se.habit.module.ServerLoaderFactory;
 import treehou.se.habit.util.Settings;
 
-public class NavigationDrawerFragment extends Fragment {
+public class NavigationDrawerFragment extends RxFragment {
 
     public static final String TAG = "NavigationDrawerFragment";
 
@@ -60,17 +66,20 @@ public class NavigationDrawerFragment extends Fragment {
     private ActionBarDrawerToggle mDrawerToggle;
 
     private DrawerLayout mDrawerLayout;
-    private ListView mDrawerListView;
+    private RecyclerView mDrawerListView;
     private View mFragmentContainerView;
 
     private int mCurrentSelectedPosition = 0;
     private boolean mUserLearnedDrawer;
 
-    private ArrayList<DrawerItem> items = new ArrayList<>();
-    private ArrayAdapter<DrawerItem> menuAdapter;
+    private List<DrawerItem> items = new ArrayList<>();
+    private List<OHSitemap> sitemaps = new ArrayList<>();
+    private DrawerAdapter menuAdapter;
 
     @Inject SharedPreferences sharedPreferences;
     @Inject Settings settings;
+    @Inject
+    ServerLoaderFactory serverLoaderFactory;
 
     public NavigationDrawerFragment() {
     }
@@ -94,11 +103,36 @@ public class NavigationDrawerFragment extends Fragment {
         items.add(new DrawerItem(getActivity().getString(R.string.servers), R.drawable.menu_servers, NavigationItems.ITEM_SERVER));
         items.add(new DrawerItem(getActivity().getString(R.string.settings), R.drawable.menu_settings, NavigationItems.ITEM_SETTINGS));
 
-        menuAdapter = new DrawerAdapter(getActivity(), items);
+        menuAdapter = new DrawerAdapter(getContext());
+        menuAdapter.setItemClickListener(item -> selectItem(item.getValue()));
+        menuAdapter.setSitemapsClickListener(this::selectSitemap);
+        menuAdapter.add(items);
     }
 
     protected ApplicationComponent getApplicationComponent() {
         return ((HabitApplication) getActivity().getApplication()).component();
+    }
+
+    /**
+     * Load servers from database and request their sitemaps.
+     */
+    private void loadSitemapsFromServers(){
+        sitemaps.clear();
+        menuAdapter.clearSitemaps();
+
+        Realm.getDefaultInstance().asObservable().compose(serverLoaderFactory.loadServersRx())
+                .compose(serverLoaderFactory.serverToSitemap(getActivity()))
+                .observeOn(AndroidSchedulers.mainThread())
+                .compose(RxLifecycle.bindFragment(this.lifecycle()))
+                .compose(serverLoaderFactory.filterDisplaySitemaps())
+                .subscribe(serverSitemaps -> {
+                    List<OHSitemap> sitemaps = serverSitemaps.second;
+
+                    boolean autoloadLast = settings.getAutoloadSitemapRx().get();
+                    this.sitemaps.addAll(sitemaps);
+
+                    menuAdapter.addSitemaps(sitemaps);
+                });
     }
 
     @Override
@@ -111,19 +145,21 @@ public class NavigationDrawerFragment extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
             Bundle savedInstanceState) {
-        mDrawerListView = (ListView) inflater.inflate(
+        mDrawerListView = (RecyclerView) inflater.inflate(
                 R.layout.fragment_navigation_drawer, container, false);
-        mDrawerListView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-                DrawerItem item = (DrawerItem) parent.getItemAtPosition(position);
-                selectItem(item.getValue(), position);
-            }
-        });
 
+        mDrawerListView.setItemAnimator(new DefaultItemAnimator());
+        mDrawerListView.setLayoutManager(new LinearLayoutManager(getContext()));
         mDrawerListView.setAdapter(menuAdapter);
-        mDrawerListView.setItemChecked(mCurrentSelectedPosition, true);
+
         return mDrawerListView;
+    }
+
+    @Override
+    public void onResume() {
+        super.onResume();
+
+        loadSitemapsFromServers();
     }
 
     public boolean isDrawerOpen() {
@@ -181,25 +217,21 @@ public class NavigationDrawerFragment extends Fragment {
         };
 
         // Defer code dependent on restoration of previous instance state.
-        mDrawerLayout.post(new Runnable() {
-            @Override
-            public void run() {
-                mDrawerToggle.syncState();
-            }
-        });
-
+        mDrawerLayout.post(() -> mDrawerToggle.syncState());
         mDrawerLayout.setDrawerListener(mDrawerToggle);
     }
 
-    private void selectItem(int value, int position) {
-        if (mDrawerListView != null) {
-            mDrawerListView.setItemChecked(position, true);
-        }
-        if (mDrawerLayout != null) {
-            mDrawerLayout.closeDrawer(mFragmentContainerView);
-        }
+    private void selectItem(int value) {
+        mDrawerLayout.closeDrawer(mFragmentContainerView);
         if (mCallbacks != null) {
             mCallbacks.onNavigationDrawerItemSelected(value);
+        }
+    }
+
+    private void selectSitemap(OHSitemap sitemap) {
+        mDrawerLayout.closeDrawer(mFragmentContainerView);
+        if (mCallbacks != null) {
+            mCallbacks.onSitemapItemSelected(sitemap);
         }
     }
 
@@ -266,13 +298,14 @@ public class NavigationDrawerFragment extends Fragment {
          * Called when an item in the navigation drawer is selected.
          */
         void onNavigationDrawerItemSelected(int value);
+        void onSitemapItemSelected(OHSitemap sitemap);
     }
 
     public static class DrawerItem{
 
         private String name;
         private int resource;
-        private int value;
+        @NavigationItems int value;
 
         public DrawerItem(String name, int resource, int value) {
             this.name = name;
@@ -288,11 +321,11 @@ public class NavigationDrawerFragment extends Fragment {
             this.name = name;
         }
 
-        public int getValue() {
+        public @NavigationItems int getValue() {
             return value;
         }
 
-        public void setValue(int value) {
+        public void setValue(@NavigationItems int value) {
             this.value = value;
         }
 
@@ -310,29 +343,223 @@ public class NavigationDrawerFragment extends Fragment {
         }
     }
 
-    class DrawerAdapter extends ArrayAdapter<DrawerItem>{
+    static class DrawerAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder> {
 
-        DrawerAdapter(Context context, List<DrawerItem> objects) {
-            super(context, R.layout.item_drawer, objects);
+        private static final int VIEW_TYPE_ITEM = 1;
+        private static final int VIEW_TYPE_SITEMAP = 2;
+
+        private List<DrawerItem> items = new ArrayList<>();
+        private List<OHSitemap> sitemaps = new ArrayList<>();
+        private Context context;
+        private OnItemClickListener itemClickListener;
+        private OnSitemapClickListener sitemapItemClickListener;
+
+        interface  OnItemClickListener{
+            void onClickItem(DrawerItem item);
+        }
+
+        interface OnSitemapClickListener {
+            void onClickItem(OHSitemap item);
+        }
+
+        static class DrawerItemHolder extends RecyclerView.ViewHolder {
+
+            private ImageView imgIcon;
+            private TextView lblName;
+
+            public DrawerItemHolder(View itemView) {
+                super(itemView);
+                imgIcon = (ImageView) itemView.findViewById(R.id.img_icon);
+                lblName = (TextView) itemView.findViewById(R.id.lbl_name);
+            }
+
+            public void update(DrawerItem entry){
+                lblName.setText(entry.getName());
+                if(entry.getResource() != 0) {
+                    imgIcon.setImageResource(entry.getResource());
+                }
+            }
+        }
+
+        static class SitemapItemHolder extends RecyclerView.ViewHolder {
+
+            private TextView lblName;
+
+            public SitemapItemHolder(View itemView) {
+                super(itemView);
+                lblName = (TextView) itemView.findViewById(R.id.lbl_sitemap);
+            }
+
+            public void update(OHSitemap entry){
+                lblName.setText(entry.getName());
+            }
+        }
+
+        public DrawerAdapter(Context context) {
+            this.context = context;
+        }
+
+        public void setItemClickListener(OnItemClickListener itemClickListener) {
+            this.itemClickListener = itemClickListener;
+        }
+
+        public void setSitemapsClickListener(OnSitemapClickListener itemClickListener) {
+            sitemapItemClickListener = itemClickListener;
+        }
+
+        /**
+         * Add menu items to add.
+         * @param drawerItems the items to add.
+         */
+        public void add(List<DrawerItem> drawerItems){
+            items.addAll(drawerItems);
+            notifyItemRangeInserted(0, drawerItems.size());
+        }
+
+        /**
+         * Add sitemaps that should be shown in menu.
+         *
+         * @param sitemaps the sitemaps that should be shown in menu.
+         */
+        public void addSitemaps(List<OHSitemap> sitemaps){
+            this.sitemaps.addAll(sitemaps);
+            notifyItemRangeInserted(findPosition(NavigationItems.ITEM_SITEMAPS), sitemaps.size());
+        }
+
+        /**
+         * Remove all sitemaps added to menu.
+         */
+        public void clearSitemaps(){
+            int sitemapSize = sitemaps.size();
+            notifyItemRangeRemoved(findPosition(NavigationItems.ITEM_SITEMAPS), sitemapSize);
+            sitemaps.clear();
+        }
+
+        /**
+         * Get sitemap from position.
+         * @param position the menu position to get sitemap for.
+         * @return sitemap at position.
+         */
+        private OHSitemap getSitemap(int position){
+            return sitemaps.get(position-(findPosition(NavigationItems.ITEM_SITEMAPS)+1));
+        }
+
+        /**
+         * Get sitemap from position.
+         * @param position the menu position to get sitemap for.
+         * @return sitemap at position.
+         */
+        private DrawerItem getMenuItem(int position){
+            int mainItemPosition = 0;
+            for(int i=0; i<getItemCount(); i++){
+                int itemViewType = getItemViewType(i);
+
+                if(VIEW_TYPE_ITEM == itemViewType){
+                    DrawerItem item = items.get(mainItemPosition);
+                    if(position == i) {
+                        return item;
+                    }
+                    mainItemPosition++;
+                }
+            }
+            return null;
+        }
+
+        /**
+         * Get menu item position.
+         *
+         * @param navigationItem the navigation item to search for.
+         * @return navigation item position.
+         */
+        private int findPosition(@NavigationItems int navigationItem){
+            int mainItemPosition = 0;
+            for(int i=0; i<getItemCount(); i++){
+                int itemViewType = getItemViewType(i);
+
+                if(VIEW_TYPE_ITEM == itemViewType){
+                    DrawerItem item = items.get(mainItemPosition);
+                    if(item.getValue() == navigationItem){
+                        return i;
+                    }
+                    mainItemPosition++;
+                }
+            }
+            return -1;
         }
 
         @Override
-        public View getView(int position, View convertView, ViewGroup parent) {
+        public RecyclerView.ViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
 
-            DrawerItem entry = getItem(position);
+            LayoutInflater inflater = LayoutInflater.from(context);
 
-            LayoutInflater inflater = LayoutInflater.from(getContext());
-
-            View itemView = inflater.inflate(R.layout.item_drawer, null);
-            ImageView imgIcon = (ImageView) itemView.findViewById(R.id.img_icon);
-            TextView lblName = (TextView) itemView.findViewById(R.id.lbl_name);
-
-            lblName.setText(entry.getName());
-            if(entry.getResource() != 0) {
-                imgIcon.setImageResource(entry.getResource());
+            RecyclerView.ViewHolder drawerItemHolder;
+            if(VIEW_TYPE_SITEMAP == viewType){
+                View itemView = inflater.inflate(R.layout.item_sitemap_small, null);
+                drawerItemHolder = new SitemapItemHolder(itemView);
+            }
+            else {
+                View itemView = inflater.inflate(R.layout.item_drawer, null);
+                drawerItemHolder = new DrawerItemHolder(itemView);
             }
 
-            return itemView;
+            return drawerItemHolder;
+        }
+
+
+
+        @Override
+        public void onBindViewHolder(RecyclerView.ViewHolder holder, int position) {
+            int itemType = getItemViewType(position);
+
+            if(VIEW_TYPE_SITEMAP == itemType){
+                OHSitemap sitemap = getSitemap(position);
+                SitemapItemHolder sitemapItemHolder = (SitemapItemHolder) holder;
+                sitemapItemHolder.update(sitemap);
+                sitemapItemHolder.itemView.setOnClickListener(view -> {
+                    if (sitemapItemClickListener != null) {
+                        sitemapItemClickListener.onClickItem(sitemap);
+                    }
+                });
+            }
+            else {
+                DrawerItem drawerItem = getMenuItem(position);
+                DrawerItemHolder drawerItemHolder = (DrawerItemHolder) holder;
+                drawerItemHolder.update(drawerItem);
+                drawerItemHolder.itemView.setOnClickListener(view -> {
+                    if (itemClickListener != null) {
+                        itemClickListener.onClickItem(drawerItem);
+                    }
+                });
+            }
+        }
+
+        @Override
+        public int getItemViewType(int position) {
+
+            int menuPosition = 0;
+            for(int i=0; i<=position; i++){
+                DrawerItem menuItem = items.get(menuPosition);
+                if(i == position){
+                    return VIEW_TYPE_ITEM;
+                }
+
+                if(menuItem.getValue() == NavigationItems.ITEM_SITEMAPS){
+                    for(int sitemapPosition=0; sitemapPosition<sitemaps.size(); sitemapPosition++){
+                        i++;
+                        if(i == position){
+                            return VIEW_TYPE_SITEMAP;
+                        }
+                    }
+                }
+                menuPosition++;
+            }
+
+            return -1;
+        }
+
+        @Override
+        public int getItemCount() {
+            return items.size() + sitemaps.size();
         }
     }
 }