+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 ## 5.1.0
 
 ### Enhancements
+* Show available sitemaps in menu.
 * Relaunch controller notifications on startup.
 * Show item name when launching slider activity.
 
 ### Bug Fixes
+* Close navigation drawer on back pressed.
 * Hide text icon, it os sent out by default from server when no icon selected.